Add 00000 to the PLMN list
Early during boot of the system, the modem answers with the string
"00000" to the OPERATOR request. Samsung-RIL currently drops this
response, while the proprietary RIL forwards it. Let's duplicate the
behaviour of the proprietary RIL here.
This seems to prevent that the RIL gets flooded with OPERATOR requests
during system boot.

Signed-off-by: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>